cleanup, extension function for OptionCard's modifier to navigate based on resId or enum
