--- app/src/main/java/com/example/dancognitionapp/Destination.kt
@@ -1,12 +1,0 @@
-package com.example.dancognitionapp
-
-enum class Destination {
-    StartTrial,
-    Practice,
-    AddParticipants,
-    NBack,
-    BART,
-    PreDive,
-    PostDive,
-    Dive
-}
