+++ app/src/main/java/com/example/dancognitionapp/assessment/selection/SelectionFragment.kt
@@ -17,12 +17,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.example.dancognitionapp.Destination
 import com.example.dancognitionapp.R
 import com.example.dancognitionapp.assessment.AssessmentActivity
 import com.example.dancognitionapp.ui.landing.DanCognitionTopAppBar
-import com.example.dancognitionapp.ui.landing.OptionCard
 import com.example.dancognitionapp.ui.theme.DanCognitionAppTheme
+import com.example.dancognitionapp.ui.widget.OptionCard
+import com.example.dancognitionapp.ui.widget.navigateTo
 
 class SelectionFragment: Fragment() {
     override fun onCreateView(
@@ -38,11 +38,7 @@ class SelectionFragment: Fragment() {
                     modifier = Modifier.fillMaxSize(),
                     isPractice = isPractice
                 ) { destination ->
-                    when(destination) {
-                        Destination.BART -> findNavController().navigate(R.id.bart_dest)
-                        Destination.NBack -> findNavController().navigate(R.id.nback_dest)
-                        else -> {}
-                    }
+                    findNavController().navigate(destination)
                 }
             }
         }
@@ -52,7 +48,7 @@ class SelectionFragment: Fragment() {
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun TestSelectScreen(modifier: Modifier = Modifier, isPractice: Boolean = false, onClick: (Destination) -> Unit = {}) {
+fun TestSelectScreen(modifier: Modifier = Modifier, isPractice: Boolean = false, onClick: (Int) -> Unit = {}) {
     Surface(
         modifier = Modifier.fillMaxSize(),
         color = MaterialTheme.colorScheme.background
@@ -74,7 +70,7 @@ fun TestSelectScreen(modifier: Modifier = Modifier, isPractice: Boolean = false,
 }
 
 @Composable
-fun SelectionPageContent(isPractice: Boolean, modifier: Modifier = Modifier, onClick: (Destination) -> Unit = {}) {
+fun SelectionPageContent(isPractice: Boolean, modifier: Modifier = Modifier, onClick: (Int) -> Unit = {}) {
     Column(
         verticalArrangement = Arrangement.SpaceEvenly,
         modifier = modifier
@@ -82,47 +78,45 @@ fun SelectionPageContent(isPractice: Boolean, modifier: Modifier = Modifier, onC
         if (isPractice) {
             OptionCard(
                 titleId = R.string.selection_practice_nback,
-                destination = Destination.NBack,
-                onCardClick = {
-                    onClick(it)
-                },
-                modifier = Modifier.weight(1f)
+                modifier = Modifier
+                    .weight(1f)
+                    .navigateTo(R.id.nback_dest) {
+                        onClick(it)
+                    }
             )
             OptionCard(
                 titleId = R.string.selection_practice_bart,
-                destination = Destination.BART,
-                onCardClick = {
-                    onClick(it)
-                },
-                modifier = Modifier.weight(1f)
-            )
-        } else {
-            OptionCard(
-                titleId = R.string.selection_pre_dive,
-                destination = Destination.PreDive,
-                onCardClick = {
-                    onClick(it)
-                },
                 modifier = Modifier
                     .weight(1f)
+                    .navigateTo(R.id.bart_dest) {
+                        onClick(it)
+                    }
+            )
+        } else {
+            //TODO - just BART test for now, but will be replaced with destinations that execute the full assessment
+            OptionCard(
+                titleId = R.string.selection_pre_dive,
+                modifier = Modifier
+                    .weight(1f)
+                    .navigateTo(R.id.bart_dest) {
+
+                    }
             )
             OptionCard(
                 titleId = R.string.selection_dive,
-                destination = Destination.Dive,
-                onCardClick = {
-                    onClick(it)
-                },
                 modifier = Modifier
                     .weight(1f)
+                    .navigateTo(R.id.bart_dest) {
+
+                    }
             )
             OptionCard(
                 titleId = R.string.selection_post_dive,
-                destination = Destination.PostDive,
-                onCardClick = {
-                    onClick(it)
-                },
                 modifier = Modifier
                     .weight(1f)
+                    .navigateTo(R.id.bart_dest) {
+
+                    }
             )
         }
     }
